+++ apps/E2E/src/Input/consts.ts
@@ -1,0 +1,9 @@
+export const HOMEPAGE_INPUT_BUTTON = 'Homepage_Input_Button';
+export const INPUT_TESTPAGE = 'Input_TestPage';
+
+/* E2E Testing Input 1 */
+export const INPUT_TEST_COMPONENT = 'Input_Test_Component'; // A component on each specific test page
+export const INPUT_ACCESSIBILITY_LABEL = 'E2E testing Input accessibility label';
+
+/* E2E Testing Input 2 */
+export const INPUT_NO_A11Y_LABEL_COMPONENT = 'Input_No_A11y_label_Component';