+++ packages/components/Menu/src/MenuList/MenuList.types.ts
@@ -13,7 +13,17 @@ export interface MenuListTokens extends LayoutTokens, IBackgroundColorTokens {
    * Space between items in pixels
    */
   gap?: number;
+
+  /**
+   * States of the list control
+   */
   hasMaxHeight?: MenuListTokens;
+
+  /**
+   * Corner radius of the menu list
+   * @platform android
+   */
+  cornerRadius?: number;
 }
 
 export interface MenuListProps extends Omit<IViewProps, 'onPress'> {
@@ -42,6 +52,7 @@ export interface MenuListProps extends Omit<IViewProps, 'onPress'> {
    *
    * This option is useful for programmatically generated items to provide
    * text for options that end up having tuncated text.
+   * @platform win32, macOS
    */
   hasTooltips?: boolean;