+++ packages/experimental/AppearanceAdditions/src/NativeAppearanceAdditions.ts
@@ -5,6 +5,9 @@ export const NativeAppearanceAdditions = {
   addListener: (_: string) => {},
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   removeListeners: (_: number) => {},
+  initializeTraitCollection: (_: number) => {
+    console.warn('NativeAppearanceAdditions is only available on iOS');
+  },
   horizontalSizeClass: () => {
     console.warn('NativeAppearanceAdditions is only available on iOS');
     return 'regular' as SizeClass;