+++ packages/components/Menu/CHANGELOG.md
@@ -1,8 +1,123 @@
 # Change Log - @fluentui-react-native/menu
 
-This log was last generated on Tue, 14 Mar 2023 20:50:45 GMT and should not be manually modified.
+This log was last generated on Wed, 19 Apr 2023 10:08:39 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.6.0
+
+Wed, 19 Apr 2023 10:08:39 GMT
+
+### Minor changes
+
+- add blackbox example and remove unused tokens (rohanpd.work@gmail.com)
+
+## 1.5.0
+
+Wed, 12 Apr 2023 18:42:22 GMT
+
+### Minor changes
+
+- menu radio strcutured (rohanpd.work@gmail.com)
+- Bump @fluentui-react-native/button to v0.32.50
+
+## 1.4.34
+
+Mon, 10 Apr 2023 19:48:24 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.23.12
+- Bump @fluentui-react-native/focus-zone to v0.11.41
+- Bump @fluentui-react-native/icon to v0.17.25
+- Bump @fluentui-react-native/interactive-hooks to v0.22.29
+- Bump @fluentui-react-native/text to v0.19.31
+- Bump @fluentui-react-native/button to v0.32.49
+
+## 1.4.33
+
+Fri, 24 Mar 2023 15:52:40 GMT
+
+### Patches
+
+- Fix for keyboard focus is lost when hovering out of a menu (email not defined)
+
+## 1.4.32
+
+Thu, 23 Mar 2023 20:49:23 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.23.11
+- Bump @fluentui-react-native/focus-zone to v0.11.40
+- Bump @fluentui-react-native/icon to v0.17.24
+- Bump @fluentui-react-native/interactive-hooks to v0.22.28
+- Bump @fluentui-react-native/text to v0.19.30
+- Bump @fluentui-react-native/button to v0.32.48
+
+## 1.4.31
+
+Thu, 23 Mar 2023 13:35:57 GMT
+
+### Patches
+
+- Revert n of m implementation for submenus (krsiler@microsoft.com)
+
+## 1.4.30
+
+Thu, 23 Mar 2023 06:05:12 GMT
+
+### Patches
+
+- Update Menu documentation (ruaraki@microsoft.com)
+
+## 1.4.29
+
+Tue, 21 Mar 2023 21:53:47 GMT
+
+### Patches
+
+- Trigger manual bump (no changes) (krsiler@microsoft.com)
+- Bump @fluentui-react-native/callout to v0.23.10
+- Bump @fluentui-react-native/focus-zone to v0.11.39
+- Bump @fluentui-react-native/framework to v0.9.8
+- Bump @fluentui-react-native/icon to v0.17.23
+- Bump @fluentui-react-native/interactive-hooks to v0.22.27
+- Bump @fluentui-react-native/text to v0.19.29
+- Bump @fluentui-react-native/tokens to v0.20.14
+- Bump @fluentui-react-native/button to v0.32.47
+
+## 1.4.28
+
+Sat, 18 Mar 2023 17:22:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.23.9
+- Bump @fluentui-react-native/focus-zone to v0.11.38
+- Bump @fluentui-react-native/framework to v0.9.7
+- Bump @fluentui-react-native/icon to v0.17.22
+- Bump @fluentui-react-native/interactive-hooks to v0.22.26
+- Bump @fluentui-react-native/text to v0.19.28
+- Bump @fluentui-react-native/tokens to v0.20.13
+- Bump @fluentui-react-native/button to v0.32.46
+
+## 1.4.27
+
+Fri, 17 Mar 2023 14:13:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.17.21
+- Bump @fluentui-react-native/button to v0.32.45
+
+## 1.4.26
+
+Thu, 16 Mar 2023 21:46:56 GMT
+
+### Patches
+
+- Fix n of m MenuItems (ruaraki@microsoft.com)
 
 ## 1.4.25