+++ packages/deprecated/theming-react-native/CHANGELOG.md
@@ -1,8 +1,43 @@
 # Change Log - @uifabricshared/theming-react-native
 
-This log was last generated on Tue, 14 Mar 2023 20:50:45 GMT and should not be manually modified.
+This log was last generated on Tue, 21 Mar 2023 21:53:45 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.16.10
+
+Tue, 21 Mar 2023 21:53:45 GMT
+
+### Patches
+
+- Trigger manual bump (no changes) (krsiler@microsoft.com)
+- Bump @fluentui-react-native/default-theme to v0.18.2
+- Bump @fluentui-react-native/win32-theme to v0.25.10
+
+## 0.16.9
+
+Sat, 18 Mar 2023 17:22:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.18.1
+- Bump @fluentui-react-native/win32-theme to v0.25.9
+
+## 0.16.8
+
+Fri, 17 Mar 2023 20:54:52 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/win32-theme to v0.25.8
+
+## 0.16.7
+
+Thu, 16 Mar 2023 21:46:56 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/win32-theme to v0.25.7
 
 ## 0.16.6