+++ packages/experimental/Checkbox/CHANGELOG.md
@@ -1,8 +1,51 @@
 # Change Log - @fluentui-react-native/experimental-checkbox
 
-This log was last generated on Tue, 14 Mar 2023 20:50:45 GMT and should not be manually modified.
+This log was last generated on Wed, 12 Apr 2023 18:42:22 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.51
+
+Wed, 12 Apr 2023 18:42:22 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.19.33
+
+## 0.13.50
+
+Mon, 10 Apr 2023 19:48:24 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.19.32
+
+## 0.13.49
+
+Thu, 23 Mar 2023 20:49:23 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.19.31
+
+## 0.13.48
+
+Tue, 21 Mar 2023 21:53:46 GMT
+
+### Patches
+
+- Trigger manual bump (no changes) (krsiler@microsoft.com)
+- Bump @fluentui-react-native/checkbox to v0.19.30
+- Bump @fluentui-react-native/framework to v0.9.8
+
+## 0.13.47
+
+Sat, 18 Mar 2023 17:22:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.19.29
+- Bump @fluentui-react-native/framework to v0.9.7
 
 ## 0.13.46