+++ packages/experimental/Popover/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @fluentui-react-native/popover
 
-This log was last generated on Tue, 14 Mar 2023 20:50:45 GMT and should not be manually modified.
+This log was last generated on Tue, 21 Mar 2023 21:53:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.58
+
+Tue, 21 Mar 2023 21:53:47 GMT
+
+### Patches
+
+- Trigger manual bump (no changes) (krsiler@microsoft.com)
+- Bump @fluentui-react-native/framework to v0.9.8
+
+## 0.1.57
+
+Sat, 18 Mar 2023 17:22:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.9.7
 
 ## 0.1.56