+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Tue, 14 Mar 2023 20:50:45 GMT and should not be manually modified.
+This log was last generated on Mon, 10 Apr 2023 19:48:23 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.22.29
+
+Mon, 10 Apr 2023 19:48:23 GMT
+
+### Patches
+
+- align-deps: Fix warnigns and issues (sanajmi@microsoft.com)
+
+## 0.22.28
+
+Thu, 23 Mar 2023 20:49:23 GMT
+
+### Patches
+
+- Change modifier key handling (ruaraki@microsoft.com)
+
+## 0.22.27
+
+Tue, 21 Mar 2023 21:53:47 GMT
+
+### Patches
+
+- Trigger manual bump (no changes) (krsiler@microsoft.com)
+- Bump @fluentui-react-native/framework to v0.9.8
+
+## 0.22.26
+
+Sat, 18 Mar 2023 17:22:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.9.7
 
 ## 0.22.25